moved chat to Text comp
